refactor: separate GQL and REST client
